requirements.txt
django
Pillow
django-widget-tweaks
razorpay